% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bit_color_def.R
\docType{data}
\name{bit_all_pal_graphic}
\alias{bit_all_pal_graphic}
\title{Look-up list of palettes}
\format{
An object of class \code{character} of length 7.
}
\usage{
bit_all_pal_graphic
}
\description{
Look-up list of palettes
}
\keyword{datasets}
