% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bit_color_def.R
\docType{data}
\name{bit_black}
\alias{bit_black}
\title{dpxR color black}
\format{
An object of class \code{character} of length 1.
}
\usage{
bit_black
}
\description{
\code{bit_black} provides the preferred color of black

See \url{https://media.unhcr.org/archive/Brand-Book-for-external-partners-2CZ7A2P0CGAY.html} for details.
}
\examples{
{
bit_black  # "#000000"
}
}
\seealso{
\code{\link{bit_pal_grey}} for the corresponding color palette;
\code{\link{bit_pal}} for the dpxR default color palette with 5 main colors of \code{\link{bit_pal_blue}};
\code{\link{bit_pal_primary}} for a dpxR color palette with primary colors;
\code{\link{seecol}} to show and use color palettes.

Other preferred colors: 
\code{\link{bit_blue}},
\code{\link{bit_darkblue}},
\code{\link{bit_green}},
\code{\link{bit_grey}},
\code{\link{bit_red}},
\code{\link{bit_violet}},
\code{\link{bit_white}},
\code{\link{bit_yellow}}
}
\concept{preferred colors}
\keyword{datasets}
