% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bit_color_def.R
\docType{data}
\name{bit_blue}
\alias{bit_blue}
\title{bitR color blue}
\format{
An object of class \code{character} of length 1.
}
\usage{
bit_blue
}
\description{
Preferred color blue (as HEX character value): ----
\code{bit_blue} provides the preferred color of \code{\link{bit_pal_blue}}
(as an atomic HEX character value) and is defined as
\code{\link{bit_pal_blue}[[5]]}.

See \url{https://media.unhcr.org/archive/Brand-Book-for-external-partners-2CZ7A2P0CGAY.html} for details.
}
\examples{
{
bit_blue  # HEX character "#0072BC" (as value)
all.equal(bit_blue, bit_pal_blue[[5]])  # TRUE (same HEX values)

seecol(bit_blue)  # view color and details
}
}
\seealso{
\code{\link{bit_pal_blue}} for the corresponding color palette;
\code{\link{bit_pal}} for the bitR default color palette with 5 main colors of \code{\link{bit_pal_blue}};
\code{\link{bit_pal_primary}} for a bitR color palette with primary colors;
\code{\link{seecol}} to show color palettes;
\code{\link{usecol}} to use color palettes.

Other preferred colors: 
\code{\link{bit_black}},
\code{\link{bit_darkblue}},
\code{\link{bit_green}},
\code{\link{bit_grey}},
\code{\link{bit_red}},
\code{\link{bit_white}},
\code{\link{bit_yellow}}
}
\concept{preferred colors}
\keyword{datasets}
