% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bit_ggplot_theme.R
\name{bit_discrete}
\alias{bit_discrete}
\title{Ggplot scale color discrete and continuous palette based on the bit_pal_graphic}
\usage{
bit_discrete()
}
\description{
Ggplot scale color discrete and continuous palette based on the bit_pal_graphic
}
\examples{
{
library(scales)
## to fix: the below...
#scales::show_col(bit_discrete(9))
}
}
\author{
ipda | unhcr.org | 2020
}
