% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bit_color_def.R
\docType{data}
\name{bit_grey}
\alias{bit_grey}
\title{bitR color grey}
\format{
An object of class \code{character} of length 1.
}
\usage{
bit_grey
}
\description{
\code{bit_grey} provides the preferred color of \code{\link{bit_pal_grey}}
(as an atomic HEX character value) and is defined as
\code{\link{bit_pal_grey}[[4]]}.

See \url{https://media.unhcr.org/archive/Brand-Book-for-external-partners-2CZ7A2P0CGAY.html} for details.
}
\examples{
{
bit_grey  # HEX character "#666666" (as value)
all.equal(bit_grey, bit_pal_grey[[4]])  # TRUE (same HEX values)

seecol(bit_grey)  # view color and details
}
}
\seealso{
\code{\link{bit_pal_grey}} for the corresponding color palette;
\code{\link{bit_pal}} for the bitR default color palette with 5 main colors of \code{\link{bit_pal_blue}};
\code{\link{bit_pal_primary}} for a bitR color palette with primary colors;
\code{\link{seecol}} to show and use color palettes.

Other preferred colors: 
\code{\link{bit_black}},
\code{\link{bit_blue}},
\code{\link{bit_darkblue}},
\code{\link{bit_green}},
\code{\link{bit_red}},
\code{\link{bit_white}},
\code{\link{bit_yellow}}
}
\concept{preferred colors}
\keyword{datasets}
