% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bit_color_def.R
\docType{data}
\name{bit_orange}
\alias{bit_orange}
\title{bitR color orange}
\format{
An object of class \code{character} of length 1.
}
\usage{
bit_orange
}
\description{
기본 오렌지 색상 (HEX 값으로): #F37321
\code{bit_orange}는 \code{\link{bit_pal_orange}}의 기본 색상을 제공하며,
(HEX 문자값 형식의) \code{\link{bit_pal_orange}[[5]]}로 정의됩니다.
}
\examples{
{
bit_orange  # HEX 문자 형식: "#F37321"
all.equal(bit_orange, bit_pal_orange[[5]])  # TRUE (HEX값과 동일)

seecol(bit_orange)  # 색상을 조회
}
}
\seealso{
\code{\link{bit_pal_orange}} bitR 색상 팔레트 대체 색상의 경우;
\code{\link{bit_pal}} 5가지 주요 색상이 있는 bitR 기본 색상 팔레트. \code{\link{bit_pal_orange}};
\code{\link{bit_pal_primary}} 기본 색상이 있는 bitR 색상 팔레트의 경우;
\code{\link{seecol}} 팔래트 색상 보기;
\code{\link{usecol}} 팔레트 색상 사용하기.
}
\concept{선호 색상}
\keyword{datasets}
