% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bit_color_def.R
\docType{data}
\name{bit_pal}
\alias{bit_pal}
\title{bit default (web/sRGB) Primary color palette}
\format{
An object of class \code{data.frame} with 1 rows and 11 columns.
}
\usage{
bit_pal
}
\description{
bitR primary color palette. \code{bit_pal}
provides the default bitR color palette
as a data frame containing 11 colors.

This is the primary (web/sRGB) scale.

See \url{https://media.unhcr.org/archive/Brand-Book-for-external-partners-2CZ7A2P0CGAY.html} for details.
}
\examples{
{
bit_pal
dim(bit_pal)  # 1 11

# Access by position:
bit_pal[1]    # 1nd named color "blue5" (as df)
bit_pal[[1]]  # 2nd color value "#0072BC"

# Access by name:
bit_pal["blue5"]    # color "blue5" (as df)
bit_pal[["blue5"]]  # color value "#0072BC"

# Plotting palette:
seecol(bit_pal)
}

}
\seealso{
\code{\link{bit_pal_blue}} for the bitR default blue palette;
\code{\link{bit_pal_primary}} for a bitR color palette with primary colors;
\code{\link{seecol}} to show color palettes;
\code{\link{usecol}} to use color palettes.

Other color palettes: 
\code{\link{bit_pal_blue}},
\code{\link{bit_pal_darkblue}},
\code{\link{bit_pal_graphic}},
\code{\link{bit_pal_green}},
\code{\link{bit_pal_grey}},
\code{\link{bit_pal_orange}},
\code{\link{bit_pal_primary}},
\code{\link{bit_pal_red}},
\code{\link{bit_pal_yellow}}
}
\concept{color palettes}
\keyword{datasets}
