% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bit_color_def.R
\docType{data}
\name{bit_pal_darkblue}
\alias{bit_pal_darkblue}
\title{dpxR darkblue color palette.}
\format{
An object of class \code{data.frame} with 1 rows and 6 columns.
}
\usage{
bit_pal_darkblue
}
\description{
darkblue
 CMYK 99 83 36 27
 RGB 24 55 95
 HEX #18375F
\code{bit_pal_darkblue} provides an additional dpxR color palette
as a data frame containing 6 colors (shades of
\code{\link{bit_darkblue}}).

See \url{https://media.unhcr.org/archive/Data-visualisation-colour-palette-2CZ7A2HOVFP9.html} for details.
}
\examples{
{
bit_pal_darkblue
dim(bit_pal_darkblue)  # 1 6
bit_pal_darkblue[6]    # preferred (named) color "darkblue6"
bit_pal_darkblue[[6]]  # preferred color "darkblue6" OR "#18375F"

# Plotting palette:
seecol(bit_pal_darkblue)
}
}
\seealso{
\code{\link{bit_pal}} for the dpxR default color palette with all 5 colors of \code{\link{bit_pal_blue}};
\code{\link{bit_pal_blue}} for an alternative blue dpxR color palette;
\code{\link{bit_pal_primary}} for a dpxR color palette with primary colors;
\code{\link{seecol}} to show color palettes;
\code{\link{usecol}} to use color palettes.

Other color palettes: 
\code{\link{bit_pal_blue}},
\code{\link{bit_pal_graphic}},
\code{\link{bit_pal_green}},
\code{\link{bit_pal_grey}},
\code{\link{bit_pal_orange}},
\code{\link{bit_pal_primary}},
\code{\link{bit_pal_red}},
\code{\link{bit_pal_violet}},
\code{\link{bit_pal_yellow}},
\code{\link{bit_pal}}
}
\concept{color palettes}
\keyword{datasets}
