% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bit_color_def.R
\docType{data}
\name{bit_pal_graphic}
\alias{bit_pal_graphic}
\title{Scale of all XX preferred colors for graphic creation}
\format{
An object of class \code{data.frame} with 1 rows and 7 columns.
}
\usage{
bit_pal_graphic
}
\description{
dpxR preferred graphic colors to be use as a main set for graphics creation.
\code{bit_pal_graphic} provides an additional dpxR color palette
that collects the preferred dpxR colors for graphics as a data frame containing 6 colors.

See \url{https://media.unhcr.org/archive/Data-visualisation-colour-palette-2CZ7A2HOVFP9.html} for details.
}
\examples{
{
bit_pal_graphic
dim(bit_pal_graphic)

# Access by position:
bit_pal_graphic[1]    # color blue (as df)
bit_pal_graphic[[1]]  # color value "#0072BC"

# Access by name:
bit_pal_graphic["Blue"]   # color "blue5" (as df)
bit_pal_graphic[["Blue"]] # color value "#0072BC"

# Plotting palette:
seecol(bit_pal_graphic)
}

}
\seealso{
\code{\link{bit_pal}} for the default dpxR color palette;
\code{\link{seecol}} to show color palettes;
\code{\link{usecol}} to use color palettes.

Other color palettes: 
\code{\link{bit_pal_blue}},
\code{\link{bit_pal_darkblue}},
\code{\link{bit_pal_green}},
\code{\link{bit_pal_grey}},
\code{\link{bit_pal_orange}},
\code{\link{bit_pal_primary}},
\code{\link{bit_pal_red}},
\code{\link{bit_pal_violet}},
\code{\link{bit_pal_yellow}},
\code{\link{bit_pal}}
}
\concept{color palettes}
\keyword{datasets}
