% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bit_color_def.R
\docType{data}
\name{bit_pal_green}
\alias{bit_pal_green}
\title{dpxR greem color palette.}
\format{
An object of class \code{data.frame} with 1 rows and 6 columns.
}
\usage{
bit_pal_green
}
\description{
dpxR green color palette.
GREEN:  CMYK 90 0 52 0 / RGB 0 179 152 / HEX #00B398
\code{pal_green} provides an additional dpxR color palette
as a data frame containing 6 colors (shades of
\code{\link{green}}).

See \url{https://media.unhcr.org/archive/Data-visualisation-colour-palette-2CZ7A2HOVFP9.html} for details.
}
\examples{
{
bit_pal_green
dim(bit_pal_green)  # 1 6
bit_pal_green[5]    # preferred (named) color "green5"
bit_pal_green[[5]]  # preferred color "green5" OR "#00B398"

# Plotting palette:
seecol(bit_pal_green)
}
}
\seealso{
\code{\link{bit_pal}} for the dpxR default color palette with all 5 colors of \code{\link{bit_pal_blue}};
\code{\link{bit_pal_primary}} for a dpxR color palette with primary colors;
\code{\link{seecol}} to show color palettes;
\code{\link{usecol}} to use color palettes.

Other color palettes: 
\code{\link{bit_pal_blue}},
\code{\link{bit_pal_darkblue}},
\code{\link{bit_pal_graphic}},
\code{\link{bit_pal_grey}},
\code{\link{bit_pal_orange}},
\code{\link{bit_pal_primary}},
\code{\link{bit_pal_red}},
\code{\link{bit_pal_violet}},
\code{\link{bit_pal_yellow}},
\code{\link{bit_pal}}
}
\concept{color palettes}
\keyword{datasets}
