% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bit_color_def.R
\docType{data}
\name{bit_pal_grey}
\alias{bit_pal_grey}
\title{bitR grey color palette}
\format{
An object of class \code{data.frame} with 1 rows and 6 columns.
}
\usage{
bit_pal_grey
}
\description{
Using a Rich black for text can make printing
 difficult and the text less sharp. We recommend body copy
 is set in C0 M0 Y0 K100. \code{pal_grey} provides an additional bitR color palette
as a data frame containing 6 colors (shades of \code{\link{black}}).

See \url{https://media.unhcr.org/archive/Brand-Book-for-external-partners-2CZ7A2P0CGAY.html} for details.
}
\examples{
{
bit_pal_grey
dim(bit_pal_grey)  # 1 6
bit_pal_grey[4]    # preferred (named) color "grey4"
bit_pal_grey[[4]]  # preferred color "grey4" OR "#666666"

# Plotting palette:
seecol(bit_pal_grey)
}
}
\seealso{
\code{\link{bit_pal}} for the bitR default color palette with 5 main colors of \code{\link{bit_pal_blue}};
\code{\link{bit_pal_darkblue}} for alternative to grey bitR color palettes;
\code{\link{bit_pal_primary}} for a bitR color palette with primary colors;
\code{\link{seecol}} to show color palettes;
\code{\link{usecol}} to use color palettes.

Other color palettes: 
\code{\link{bit_pal_blue}},
\code{\link{bit_pal_darkblue}},
\code{\link{bit_pal_graphic}},
\code{\link{bit_pal_green}},
\code{\link{bit_pal_orange}},
\code{\link{bit_pal_primary}},
\code{\link{bit_pal_red}},
\code{\link{bit_pal_yellow}},
\code{\link{bit_pal}}
}
\concept{color palettes}
\keyword{datasets}
