% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bit_color_def.R
\docType{data}
\name{bit_pal_primary}
\alias{bit_pal_primary}
\title{Scale of all 4 preferred colors, Other combinations of bitR color palettes}
\format{
An object of class \code{data.frame} with 1 rows and 4 columns.
}
\usage{
bit_pal_primary
}
\description{
\code{bit_pal_primary} provides an additional bitR color palette
that collects the primary bitR color as a data frame containing 4 colors.

See \url{https://media.unhcr.org/archive/Brand-Book-for-external-partners-2CZ7A2P0CGAY.html} for details.
}
\examples{
{
bit_pal_primary
dim(bit_pal_primary)  # 1 4

# Access by position:
bit_pal_primary[1]    # color blue (as df)
bit_pal_primary[[1]]  # color value "#0072BC"

# Access by name:
bit_pal_primary["Blue"]   # color "blue5" (as df)
bit_pal_primary[["Blue"]] # color value "#0072BC"

# Plotting palette:
seecol(bit_pal_primary)
}
}
\seealso{
\code{\link{bit_pal}} for the default bitR color palette;
\code{\link{seecol}} to show color palettes;
\code{\link{usecol}} to use color palettes.

Other color palettes: 
\code{\link{bit_pal_blue}},
\code{\link{bit_pal_darkblue}},
\code{\link{bit_pal_graphic}},
\code{\link{bit_pal_green}},
\code{\link{bit_pal_grey}},
\code{\link{bit_pal_orange}},
\code{\link{bit_pal_red}},
\code{\link{bit_pal_yellow}},
\code{\link{bit_pal}}
}
\concept{color palettes}
\keyword{datasets}
