% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bit_color_def.R
\docType{data}
\name{bit_pal_red}
\alias{bit_pal_red}
\title{dpxR red color palette.}
\format{
An object of class \code{data.frame} with 1 rows and 6 columns.
}
\usage{
bit_pal_red
}
\description{
dpxR red color palette.
 RED: CMYK 0 87 52 0 / RGB 239 74 96 / HEX #EF4A60

\code{pal_red} provides an additional dpxR color palette
as a data frame containing 6 colors (shades of
\code{\link{bit_red}}).

See \url{https://media.unhcr.org/archive/Data-visualisation-colour-palette-2CZ7A2HOVFP9.html} for details.
}
\examples{
{
bit_pal_red
dim(bit_pal_red)  # 1 6
bit_pal_red[5]    # preferred (named) color "red5"
bit_pal_red[[5]]  # preferred color "red5" OR "#EF4A60"

# Plotting palette:
seecol(bit_pal_red)
}
}
\seealso{
\code{\link{bit_pal}} for the dpxR default color palette with all 5 colors of \code{\link{bit_pal_blue}};
\code{\link{bit_pal_primary}} for a dpxR color palette with primary colors;
\code{\link{seecol}} to show color palettes;
\code{\link{usecol}} to use color palettes.

Other color palettes: 
\code{\link{bit_pal_blue}},
\code{\link{bit_pal_darkblue}},
\code{\link{bit_pal_graphic}},
\code{\link{bit_pal_green}},
\code{\link{bit_pal_grey}},
\code{\link{bit_pal_orange}},
\code{\link{bit_pal_primary}},
\code{\link{bit_pal_violet}},
\code{\link{bit_pal_yellow}},
\code{\link{bit_pal}}
}
\concept{color palettes}
\keyword{datasets}
