% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bit_color_def.R
\docType{data}
\name{bit_pal_yellow}
\alias{bit_pal_yellow}
\title{dpxR yellow color palette.}
\format{
An object of class \code{data.frame} with 1 rows and 6 columns.
}
\usage{
bit_pal_yellow
}
\description{
The secondary accent colour may be used in some circumstances
for a graphic element.
YELLOW: PANTONE PROCESS YELLOW C/U
  CMYK 3 1 100 0
  RGB 250 235 0
  HEX #FAEB00
  \code{bit_pal_yellow} provides an additional dpxR color palette
as a data frame containing 6 colors (shades of
\code{\link{bit_yellow}}).

See \url{https://media.unhcr.org/archive/Brand-Book-for-external-partners-2CZ7A2P0CGAY.html} for details.
}
\examples{
{
bit_pal_yellow
dim(bit_pal_yellow)  # 1 6
bit_pal_yellow[5]    # preferred (named) color "yellow5"
bit_pal_yellow[[5]]  # preferred color "yellow5" OR "#FAEB00"

# Plotting palette:
seecol(bit_pal_yellow)
}
}
\seealso{
\code{\link{bit_pal}} for the dpxR default color palette with all 5 colors of \code{\link{bit_yellow}};
\code{\link{bit_pal_primary}} for a dpxR color palette with primary colors;
\code{\link{seecol}} to show color palettes;
\code{\link{usecol}} to use color palettes.

Other color palettes: 
\code{\link{bit_pal_blue}},
\code{\link{bit_pal_darkblue}},
\code{\link{bit_pal_graphic}},
\code{\link{bit_pal_green}},
\code{\link{bit_pal_grey}},
\code{\link{bit_pal_orange}},
\code{\link{bit_pal_primary}},
\code{\link{bit_pal_red}},
\code{\link{bit_pal_violet}},
\code{\link{bit_pal}}
}
\concept{color palettes}
\keyword{datasets}
