% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bit_color_def.R
\docType{data}
\name{bit_red}
\alias{bit_red}
\title{Preferred color red (as HEX character value)}
\format{
An object of class \code{character} of length 1.
}
\usage{
bit_red
}
\description{
dpxR color red.

\code{red} provides the preferred color of \code{\link{bit_pal_red}}
(as an atomic HEX character value) and is defined as
\code{\link{bit_pal_red}[[5]]}.

See \url{https://media.unhcr.org/archive/Data-visualisation-colour-palette-2CZ7A2HOVFP9.html} for details.
}
\examples{
{
bit_red  # HEX character "#EF4A60" (as value)
all.equal(bit_red, bit_pal_red[[5]])  # TRUE (same HEX values)
seecol(bit_red)  # view color and details
}
}
\seealso{
\code{\link{bit_pal_red}} for the corresponding color palette;
\code{\link{bit_pal}} for the dpxR default color palette with all 5 colors of \code{\link{bit_pal_blue}};
\code{\link{bit_pal_primary}} for a dpxR color palette with primary colors;
\code{\link{seecol}} to show color palettes;
\code{\link{usecol}} to use color palettes.

Other preferred colors: 
\code{\link{bit_black}},
\code{\link{bit_blue}},
\code{\link{bit_darkblue}},
\code{\link{bit_green}},
\code{\link{bit_grey}},
\code{\link{bit_violet}},
\code{\link{bit_white}},
\code{\link{bit_yellow}}
}
\concept{preferred colors}
\keyword{datasets}
