% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bit_ggplot_theme.R
\name{bit_scale_fill_discrete}
\alias{bit_scale_fill_discrete}
\alias{bit_scale_colour_discrete}
\title{Discrete  fill scales based on the bitR graphic palette}
\usage{
bit_scale_fill_discrete(...)
}
\description{
See [bit_discrete()].
}
\author{
ipda | unhcr.org | 2020
}
