% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bit_color_def.R
\docType{data}
\name{bit_violet}
\alias{bit_violet}
\title{dpxR color violet}
\format{
An object of class \code{character} of length 1.
}
\usage{
bit_violet
}
\description{
Preferred color violet (as HEX character value): ----
\code{bit_violet} provides the preferred color of \code{\link{bit_pal_violet}}
(as an atomic HEX character value) and is defined as
\code{\link{bit_pal_violet}[[5]]}.

See \url{https://www.schemecolor.com/violet-gradient.php} for details.
}
\examples{
{
violet  # HEX character "#552586" (as value)
all.equal(bit_violet, bit_pal_violet[[5]])  # TRUE (same HEX values)

seecol(bit_violet)  # view color and details
}
}
\seealso{
\code{\link{bit_pal_violet}} for the corresponding color palette;
\code{\link{bit_pal}} for the dpxR default color palette with 5 main colors of \code{\link{bit_pal_violet}};
\code{\link{bit_pal_primary}} for a dpxR color palette with primary colors;
\code{\link{seecol}} to show color palettes;
\code{\link{usecol}} to use color palettes.

Other preferred colors: 
\code{\link{bit_black}},
\code{\link{bit_blue}},
\code{\link{bit_darkblue}},
\code{\link{bit_green}},
\code{\link{bit_grey}},
\code{\link{bit_red}},
\code{\link{bit_white}},
\code{\link{bit_yellow}}
}
\concept{preferred colors}
\keyword{datasets}
