% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bit_color_def.R
\docType{data}
\name{bit_yellow}
\alias{bit_yellow}
\title{Preferred color yellow (as HEX character value)}
\format{
An object of class \code{character} of length 1.
}
\usage{
bit_yellow
}
\description{
dpxR color yellow.

\code{yellow} provides the preferred color of \code{\link{bit_pal_yellow}}
(as an atomic HEX character value) and is defined as
\code{\link{bit_pal_yellow}[[5]]}.

See \url{https://media.unhcr.org/archive/Brand-Book-for-external-partners-2CZ7A2P0CGAY.html} for details.
}
\examples{
{
bit_yellow  # HEX character "#FAEB00" (as value)
all.equal(bit_yellow, bit_pal_yellow[[5]])  # TRUE (same HEX values)
seecol(bit_yellow)  # view color and details
}

}
\seealso{
\code{\link{bit_pal_yellow}} for the corresponding color palette;
\code{\link{bit_pal}} for the bit default color palette with all 5 colors of \code{\link{bit_pal_yellow}};
\code{\link{bit_pal_primary}} for a dpxR color palette with primary colors;
\code{\link{seecol}} to show color palettes;
\code{\link{usecol}} to use color palettes.

Other preferred colors: 
\code{\link{bit_black}},
\code{\link{bit_blue}},
\code{\link{bit_darkblue}},
\code{\link{bit_green}},
\code{\link{bit_grey}},
\code{\link{bit_red}},
\code{\link{bit_violet}},
\code{\link{bit_white}}
}
\concept{preferred colors}
\keyword{datasets}
