% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bit_templ_paged.R
\name{bitR_orange_paged}
\alias{bitR_orange_paged}
\alias{bitR_blue_paged}
\title{Create a paged HTML document suitable for printing with bitR Brand}
\usage{
bitR_orange_paged(...)

bitR_blue_paged(...)
}
\arguments{
\item{...}{Arguments passed to \code{pagedown::\link[pagedown]{html_paged}}.}
}
\value{
An R Markdown output format.
}
\description{
Create a paged HTML document suitable for printing with bitR Brand
}
\references{
\url{https://pagedown.rbind.io}
}
