% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bitreport.R
\docType{package}
\name{bitReport}
\alias{bitReport}
\title{\code{bitReport} package}
\description{
A collection of template, palettes and themes for plots, documents, and slides
based on bitR branding
}
\details{
See the README on
\href{https://github.com/unhcr/unhcRstyle}{GitHub}
}
\references{
https://unhcr-web.github.io/unhcRstyle/docs/
}
