% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bit_color_util.R
\name{col2hex}
\alias{col2hex}
\title{col2hex color conversion function in grDevices}
\usage{
col2hex(col, alpha = alpha)
}
\description{
Utility functions to access and plot color palettes
}
\examples{
{
hex1 <- col2hex("black", alpha = 255/2)
hex2 <- col2hex("white", alpha = 255/2)
hex3 <- col2hex("gold", alpha = 255/2)
hex4 <- col2hex("steelblue", alpha = 255/2)
#seecol(pal = c(hex1, hex2, hex3, hex4), n = "all")
}
}
\author{
ipda | unhcr.org | 2020
}
