% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ga_viz.R
\name{compare_m_kpi}
\alias{compare_m_kpi}
\title{KPI 메트릭의 전월과의 성장률 비교}
\usage{
compare_m_kpi(
  mart_kpi,
  prev_month = "2021-02",
  curr_month = "2021-03",
  platform_cd = NULL,
  metrics = c("n_user", "n_session", "timeonsite", "n_newuser", "n_newsession",
    "newsession_rate"),
  bar_ratio = 1,
  page_size = 10,
  platform_band = FALSE
)
}
\arguments{
\item{mart_kpi}{data.frame or a \code{\link{tbl_df}} or a \code{\link{grouped_df}}.}

\item{prev_month}{character. 이전 월도.}

\item{curr_month}{character. 기준 월도.}

\item{platform_cd}{character. 표시할 대상의 플랫폼 코드.}

\item{bar_ratio}{numeric. 증감비율로 표현할 막대의 가중치. 기본값은 1이다.}

\item{page_size}{integer. 한 페이지에 출력할 레코드 수.}

\item{platform_band}{logical. 플랫폼 이름에 밴드를 지정할지의 여부. 기본값은 FALSE, TRUE이면, 짝수번쨰 플랫폼에 회색 밴드를 표시한다.}

\item{metric}{character. 집계할 메트릭.}
}
\description{
compare_m_kpi()은 월별로 집계된 마트의 KPI 메트릭을 전월과 비교하여, 그 증감률을 테이블로 표현한다.
}
\details{
집계 테이블은 월별로 집계된 GA KPI 마트를 이용한 시각화를 포함한 집계표로 일평균 사용자수,
일평균 세션수, 평균 세션시간, 일평균 신규 사용자수, 일평균 신규 세션수, 신규세션 비율을 출력한다.

platform_cd 종류는 다음과 같다.
\itemize{
  \item "ONS" : 온슈어
  \item "HMP" : 홈페이지
  \item "IWL" : 보험월렛
}
metric은 집계 지표로 다음과 같다.
\itemize{
  \item "n_user" : 일평균 사용자수 (명)
  \item "n_session" : 일평균 세션수 (건)
  \item "timeonsite" : 평균 세션시간 (초)
  \item "n_newuser" : 일평균 신규 사용자수 (명)
  \item "n_newsession" : 일평균 신규 세션수 (건)
  \item "newsession_rate" : 신규세션 비율 (%)
}
}
\examples{
compare_m_kpi(TB_D_KPIS)

compare_m_kpi(TB_D_KPIS, bar_ratio = 2, page_size = 18)

compare_m_kpi(TB_D_KPIS, metrics = c("n_user", "n_session"))

compare_m_kpi(TB_D_KPIS, bar_ratio = 1.5, page_size = 18, platform_band = TRUE)

}
