% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bit_exam_plot.R
\name{formatable_country_metric}
\alias{formatable_country_metric}
\title{국가 통계의 formattable 테이블 출력}
\usage{
formatable_country_metric(base_year = 2020, country = NULL, fixed_width = 100)
}
\arguments{
\item{base_year}{character. 통계 년도.}

\item{country}{character. 시각화할 국가 이름.}

\item{fixed_width}{integer. 컬럼의 너비}
}
\description{
formatable_country_metric()은 국가 통계를 formattable 패키지를 사용하여, 테이블로 표현한다.
}
\details{
집계 테이블은 시각화를 포함한 집계표로 인구, 기대수명, GDP, 1인당 GDP, 
GDP 증가율을 출력한다.

country 종류는 다음과 같다.
\itemize{
  \item "United States" : 미국
  \item "China" : 중국
  \item "Japan" : 일본
  \item "Germany" : 독일
  \item "United Kingdom" : 영국
  \item "India" : 인도
}
}
\examples{
formatable_country_metric()

formatable_country_metric(base_year = 2015)

formatable_country_metric(country = c("United States", "Germany", "Japan"))

}
