% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bit_color_util.R
\name{getpal_key}
\alias{getpal_key}
\title{Get a palette or list of palettes by keyword}
\usage{
getpal_key(pal = "all", n = "all", alpha = NA)
}
\description{
Helper function
}
\author{
ipda | unhcr.org | 2020
}
\keyword{internal}
