% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bit_color_util.R
\name{isCol}
\alias{isCol}
\title{isColl color conversion function in grDevices}
\usage{
isCol(color)
}
\description{
Helper function to detect any color (in an individual character string)
}
\examples{
{
isCol("white")
isCol(col2hex("black", alpha = 255/2))
isCol(NA)
isCol("bumblebee")
# BUT note:
isCol(col2rgb("white"))  # => FALSE FALSE FALSE
 }
}
\author{
ipda | unhcr.org | 2020
}
\keyword{internal}
