% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bit_color_util.R
\name{isHexCol}
\alias{isHexCol}
\title{isHexCol color conversion function in grDevices}
\usage{
isHexCol(color)
}
\description{
Helper function to detect HEX-colors
}
\examples{
{
isHexCol("black")
isHexCol(col2hex("black"))
isHexCol(rgb2hex(0, 0, 0))
}
}
\author{
ipda | unhcr.org | 2020
}
\keyword{internal}
