% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bit_simple_paged.R
\name{page_simpled}
\alias{page_simpled}
\title{Function for KANDLE simpled report template}
\usage{
page_simpled(
  back_html = TRUE,
  other_css = NULL,
  number_sections = FALSE,
  theme = c("violet", "green", "blue", "orange"),
  ...
)
}
\arguments{
\item{back_html}{logical. back-cover HTML including contact information}

\item{other_css}{Add extra css}

\item{number_sections}{Number section headings}

\item{theme}{Theme color}

\item{...}{Arguments passed to pagedown::html_paged}
}
\value{
An R Markdown output format.
}
\description{
Function for KANDLE simpled report template
}
\examples{
\dontrun{
# Requires pandoc

# Create an R Markdown example file
rmd_content <- "---
 title: 'Example bitReport HTML'
 subtitle: 'Paged Report'
 output: bitReport::bit_simpled
 ---

 ## Introduction

 This is an example of a bitReport-branded HTML paged report using `bitReport::bit_simpled`.
 "

# Write the R Markdown content to a file
example_file <- tempfile(fileext = ".Rmd")
writeLines(rmd_content, example_file)

# Render the R Markdown file
rmarkdown::render(example_file,
  output_format = bitReport::bit_simpled(),
  output_dir = tempdir()
)

# View the rendered HTML file
browseURL(render_file)
}
}
