% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bit_color_util.R
\name{plot_col}
\alias{plot_col}
\title{Plot a vector of colors (as circles or rectangles)}
\usage{
plot_col(
  x,
  ypos = 1,
  shape = "rect",
  xlen = 1,
  ylen = 1,
  distance = 0,
  plot.new = TRUE,
  ...
)
}
\description{
Helper function
}
\author{
ipda | unhcr.org | 2020
}
\keyword{internal}
