% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bit_exam_plot.R
\name{plot_country_trend}
\alias{plot_country_trend}
\title{국가 통계의 년간 트랜드 플롯}
\usage{
plot_country_trend(
  start_year = 1990,
  end_year = 2020,
  country = NULL,
  metric = c("gdp", "gdp_per_capita", "gdp_growth", "imports_gs", "exports_gs", "debt",
    "unemployment", "population", "population_growth", "life_expectancy")
)
}
\arguments{
\item{start_year}{character. 통계 시작 년도.}

\item{end_year}{character. 통계 종료 년도.}

\item{country}{character. 시각화할 국가 이름}

\item{metric}{character. 시각화할 메트릭.}
}
\description{
plot_country_trend()은 국가의 메트릭을 년간 시계열 플롯으로 시각화한다.
}
\details{
국가 통계의 년간 트랜드 플롯은 World Bank 데이터인 economies를 이용한 시계열 라인플롯이다.

country 종류는 다음과 같다.
\itemize{
  \item "United States" : 미국
  \item "China" : 중국
  \item "Japan" : 일본
  \item "Germany" : 독일
  \item "United Kingdom" : 영국
  \item "India" : 인도
}
metric 종류는 다음과 같다.
\itemize{
  \item "gdp" : GDP
  \item "gdp_per_capita" : 인당 GDP
  \item "gdp_growth" : GDP 성장률
  \item "imports_gs" : 상품/서비스 수입
  \item "exports_gs" : 상품/서비스 수출
  \item "debt" : 국개 채무
  \item "unemployment" : 실업률
  \item "population" : 인구
  \item "population_growth" : 인구증가율
  \item "life_expectancy" : 기대수명
}
}
\examples{
plot_country_trend()

plot_country_trend(metric = "population")

plot_country_trend(start_year = 1995, metric = "life_expectancy")

plot_country_trend(metric = "unemployment", country = "United States")
plot_country_trend(metric = "unemployment", country = c("China", "Japan"))

}
