% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ga_viz.R
\name{plot_d_kpi}
\alias{plot_d_kpi}
\title{KPI 메트릭의 일 트랜드 플롯}
\usage{
plot_d_kpi(
  mart_kpi,
  start_date = "2021-03-01",
  end_date = "2021-03-31",
  platform_cd = NULL,
  metric = c("n_user", "n_newuser", "n_session", "n_newsession", "n_screen", "n_page",
    "tot_timeonsite", "tot_timeonscreen")[1],
  moving = c("d1", "w1", "w2", "w3", "w4", "m1")[5]
)
}
\arguments{
\item{mart_kpi}{data.frame or a \code{\link{tbl_df}} or a \code{\link{grouped_df}}.}

\item{start_date}{character. 통계 시작 일자.}

\item{end_date}{character. 통계 종료 일자.}

\item{platform_cd}{character. 시각화할 플랫폼 코드.}

\item{metric}{character. 시각화할 메트릭.}

\item{moving}{character. 메트릭 집계 주기}
}
\description{
plot_d_kpi()은 일별로 집계된 마트의 KPI 메트릭을 시계열 플롯으로 시각화한다.
}
\details{
일 트랜드 플롯은 일별로 집계된 GA KPI 마트를 이용한 시계열 라인플롯이다.

platform_cd 종류는 다음과 같다.
\itemize{
  \item "ONS" : 온슈어
  \item "HMP" : 홈페이지
  \item "IWL" : 보험월렛
}
metric 종류는 다음과 같다.
\itemize{
  \item "n_user" : 활성 사용자
  \item "n_newuser" : 신규 사용자
  \item "n_session" : 세션수
  \item "n_newsession" : 신규 세션수
  \item "n_screen" : 화면수
  \item "n_page" : 페이지수
  \item "tot_timeonsite" : 세션 시간
  \item "tot_timeonscreen" : 스크린 시간
}
moving 종류는 집계 기간으로 다음과 같다.
\itemize{
  \item "d1" : 일 집계
  \item "w1" : 7일 집계
  \item "w2" : 14일 집계
  \item "w3" : 21일 집계
  \item "w4" : 28일 집계
  \item "m1" : 월 집계
}
}
\examples{
plot_d_kpi(TB_D_KPIS)

plot_d_kpi(TB_D_KPIS, metric = "n_session", moving = "w4")

plot_d_kpi(TB_D_KPIS, start_date = "2021-01-01", metric = "n_session",
           moving = "w1")

plot_d_kpi(TB_D_KPIS, metric = "n_session", platform_cd = "IWL")
plot_d_kpi(TB_D_KPIS, metric = "n_session", platform_cd = c("IWL", "HMP"))

}
