% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bit_color_util.R
\name{plot_shape}
\alias{plot_shape}
\title{Plot a shape in a certain color}
\usage{
plot_shape(
  pos_x,
  pos_y,
  col_fill,
  col_brd = NA,
  xlen = 1,
  ylen = 1,
  shape = "rect",
  ...
)
}
\description{
Helper function
}
\author{
ipda | unhcr.org | 2020
}
\keyword{internal}
