% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bit_templ_paged.R
\name{report_paged}
\alias{report_paged}
\alias{bitR_violet_paged}
\alias{bitR_green_paged}
\title{Create a paged HTML document suitable for printing with bitR Brand}
\usage{
report_paged(
  front_cover = NULL,
  other_css = NULL,
  theme = c("violet", "green", "blue", "orange"),
  ...
)

bitR_violet_paged(front_cover = NULL, ...)

bitR_green_paged(front_cover = NULL, ...)
}
\arguments{
\item{...}{Arguments passed to \code{pagedown::\link[pagedown]{html_paged}}.}
}
\value{
An R Markdown output format.
}
\description{
Create a paged HTML document suitable for printing with bitR Brand
}
\references{
\url{https://pagedown.rbind.io}
}
