% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bit_color_util.R
\name{rgb2hex}
\alias{rgb2hex}
\title{rgb2hex color conversion function in grDevices}
\usage{
rgb2hex(R, G, B)
}
\description{
Utility functions to access and plot color palettes
}
\examples{
{
rgb2hex(255, 255, 255)
rgb2hex(0, 0, 0)
}
}
\author{
ipda | unhcr.org | 2020
}
\keyword{internal}
