% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bit_exam_plot.R
\name{table_country_metric}
\alias{table_country_metric}
\title{국가 통계의 집계 테이블 출력}
\usage{
table_country_metric(base_year = 2020, country = NULL)
}
\arguments{
\item{base_year}{character. 통계 년도.}

\item{country}{character. 시각화할 국가 이름.}
}
\description{
table_country_metric()는 국가의 메트릭을 플롯을 포함한 테이블로 표현한다.
}
\details{
집계 테이블은 시각화를 포함한 집계표로 인구, 기대수명, GDP, 1인당 GDP, 
GDP 증가율을 출력한다.

country 종류는 다음과 같다.
\itemize{
  \item "United States" : 미국
  \item "China" : 중국
  \item "Japan" : 일본
  \item "Germany" : 독일
  \item "United Kingdom" : 영국
  \item "India" : 인도
}
}
\examples{
table_country_metric()

table_country_metric(base_year = 2015)

table_country_metric(country = c("United States", "Germany", "Japan"))

}
