% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ga_viz.R
\name{table_d_kpi}
\alias{table_d_kpi}
\title{KPI 메트릭의 집계 테이블 출력}
\usage{
table_d_kpi(
  mart_kpi,
  base_date = "2021-03-31",
  platform_cd = NULL,
  moving = c("d1", "w1", "w2", "w3", "w4", "m1")[5]
)
}
\arguments{
\item{mart_kpi}{data.frame or a \code{\link{tbl_df}} or a \code{\link{grouped_df}}.}

\item{base_date}{character. 통계 기준 일자.}

\item{platform_cd}{character. 표시할 대상의 플랫폼 코드.}

\item{moving}{character. 메트릭 집계 주기}
}
\description{
table_d_kpi()은 일별로 집계된 마트의 KPI 메트릭을 플롯을 포함한 테이블로 표현한다.
}
\details{
집계 테이블은 일별로 집계된 GA KPI 마트를 이용한 시각화를 포함한 집계표로 활성 사용자수,
세션수, 사용자당 세션수, 평균 세션시간, 신규세션 비율을 출력한다.

platform_cd 종류는 다음과 같다.
\itemize{
  \item "ONS" : 온슈어
  \item "HMP" : 홈페이지
  \item "IWL" : 보험월렛
  \item "LPS" : 라이프플러스
  \item "HLO" : HELLO
  \item "HRT" : 마음건강
  \item "SMP" : 스마트플래너
}
moving 종류는 집계 기간으로 다음과 같다.
\itemize{
  \item "d1" : 일 집계
  \item "w1" : 7일 집계
  \item "w2" : 14일 집계
  \item "w3" : 21일 집계
  \item "w4" : 28일 집계
  \item "m1" : 월 집계
}
}
\examples{
table_d_kpi(TB_D_KPIS)

table_d_kpi(TB_D_KPIS, moving = "d1")

table_d_kpi(TB_D_KPIS, base_date = "2021-03-01")

table_d_kpi(TB_D_KPIS, platform_cd = c("IWL", "HMP"))

}
