% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bit_ggplot_theme.R
\name{theme_foundation}
\alias{theme_foundation}
\title{Base clean ggplot theme to build from}
\usage{
theme_foundation(base_size = 12, base_family = "")
}
\description{
Base clean ggplot theme to build from
}
\author{
ipda | unhcr.org | 2020
}
