% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bit_ggplot_theme.R
\name{update_geom_font_defaults}
\alias{update_geom_font_defaults}
\title{Update matching font defaults for text geoms}
\usage{
update_geom_font_defaults(
  family = "NanumSquare",
  face = "plain",
  size = 3.5,
  color = grey(0.15, 1)
)
}
\arguments{
\item{family}{font family name}

\item{face}{font  face}

\item{size}{font size}

\item{color}{font color}
}
\description{
Updates [ggplot2::geom_label] and [ggplot2::geom_text] font defaults
}
\author{
ipda | unhcr.org | 2020
}
