% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bit_color_fun.R
\name{usecol}
\alias{usecol}
\title{Use a color palette (as is)}
\usage{
usecol(
  pal = bit_pal,
  n = "all",
  alpha = NA,
  use_names = FALSE,
  use_col_ramp = FALSE
)
}
\arguments{
\item{pal}{A color palette (as a vector of colors or color palettes).
Default: \code{pal = \link{bit_pal}}.}

\item{n}{An integer value specifying the desired number of colors from the palette.
Default: \code{n = "all"} (i.e., use all colors of a color palette).
For the palettes defined by \strong{bit}, \code{n} is set to a pre-defined selection of
colors if the desired number of colors is smaller than the available number.
For all other palettes and values of \code{n} larger than \code{length(pal)},
\code{n} compresses or extends the palette using \code{\link{colorRampPalette}}.}

\item{alpha}{A factor modifying the opacity alpha (as in \code{\link{adjustcolor}});
to a value in [0,1].
Default: \code{alpha = NA} (i.e., no modification of opacity).}

\item{use_names}{A logical value indicating whether colors should be returned as a named vector.
Default: \code{use_names = FALSE}, for compatibility with \code{ggplot}.}

\item{use_col_ramp}{A logical value specifying whether the default of using pre-selected colors
should be overridden and \code{\link{colorRampPalette}} should be used to process \code{n}.
Default: \code{use_col_ramp = FALSE}.}
}
\value{
A vector of colors (in character format).
}
\description{
\code{usecol} allows using a color or color palette \code{pal}
 (e.g., for plotting).  \code{usecol} also allows modifying and combining
  color palettes in various ways.
}
\examples{
{
usecol(pal = bit_pal, n = "all")  # default color palette
usecol(pal = bit_pal, n =  4)     # selecting n dedicated colors
usecol(pal = bit_pal, n = 20)     # extending color palette

# Mixing a new color palette:
pal_1 <- usecol(pal = c(rev(pal_blue), "white", pal_red))
seecol(pal_1)

# Mixing and extending a color palette:
pal_2 <- usecol(pal = c(rev(pal_green), "white", pal_red), n = 20)
seecol(pal_2)
}
}
\seealso{
\code{\link{seecol}} to plot color palettes;
\code{\link{bit_pal}} for the default bitR color palette.
}
\concept{color functions}
